% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bon_ch.R
\name{bon_ch}
\alias{bon_ch}
\title{bon_ch}
\usage{
bon_ch()
}
\value{
tibble with daily counts, proportions of runs complete, and forecasts
}
\description{
Reads window count data from fish passage center. Calculates predictions based on 5 and 10 year average run timing.
}
\examples{

bon_ch()

}
