% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bon_ch.R
\name{Bon_ch_fun}
\alias{Bon_ch_fun}
\title{Calculates predictions based on 5 and 10 year average run timing.}
\usage{
Bon_ch_fun(pred_date = NULL)
}
\arguments{
\item{dat}{}
}
\description{
Calculates predictions based on 5 and 10 year average run timing.
}
