% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_funs.R
\name{do_salmonForecasting_fun}
\alias{do_salmonForecasting_fun}
\title{ARIMA (salmonForecasting) model forecast}
\usage{
do_salmonForecasting_fun(data, cov_vec = c("log_cum_cnt", "cnt_by_flow"))
}
\arguments{
\item{cov_vec}{}
}
\description{
ARIMA (salmonForecasting) model forecast
}
