% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_funs.R
\name{do_sibregresr_fun}
\alias{do_sibregresr_fun}
\title{make forecast using sibregresr}
\usage{
do_sibregresr_fun(data, cov_vec = c("log_cum_cnt", "cnt_by_flow"))
}
\arguments{
\item{fish_river_ocean}{data}
}
\description{
make forecast using sibregresr
}
