% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_funs.R
\name{ema}
\alias{ema}
\title{Function to calculate exponential moving avg on a time series (x)}
\usage{
ema(x, ratio)
}
\arguments{
\item{ratio}{}
}
\description{
Function to calculate exponential moving avg on a time series (x)
}
