% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Joint_likelihood_RTMB.R
\name{fit_joint_likelihood}
\alias{fit_joint_likelihood}
\title{fit the joint likelihood model}
\usage{
fit_joint_likelihood(dat, forecast, forecast_log_sd)
}
\arguments{
\item{forecast_log_sd}{}
}
\description{
fit the joint likelihood model
}
