% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Joint_likelihood_RTMB2.R
\name{fit_joint_likelihood2}
\alias{fit_joint_likelihood2}
\title{fit the joint likelihood model}
\usage{
fit_joint_likelihood2(dat, forecast_season)
}
\arguments{
\item{forecast_log_sd}{}
}
\description{
fit the joint likelihood model
}
