% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDO.R
\name{fpc_laddersplit}
\alias{fpc_laddersplit}
\title{Get laddersplit count report from FPC (from Ben's fpcDamCounts)}
\usage{
fpc_laddersplit(
  pred_date = NULL,
  sdate = as.Date(paste0(lubridate::year(pred_date), "-08-01")),
  url = "https://www.fpc.org/adults/R_adultcoequeries_laddersplitreport_results_get"
)
}
\arguments{
\item{url}{}
}
\description{
Get laddersplit count report from FPC (from Ben's fpcDamCounts)
}
