% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_funs.R
\name{get_CBR_DART_flow_temp_data}
\alias{get_CBR_DART_flow_temp_data}
\title{get flow data from CRB DART}
\usage{
get_CBR_DART_flow_temp_data(forecastdate, start_year = 1960)
}
\arguments{
\item{start_year}{}
}
\description{
get flow data from CRB DART
}
