% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{get_default_model_result_path}
\alias{get_default_model_result_path}
\title{Get default local storage path for model result data}
\usage{
get_default_model_result_path()
}
\value{
Full path to the default flow data CSV
}
\description{
Get default local storage path for model result data
}
