% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{get_default_ocean_path}
\alias{get_default_ocean_path}
\title{Get default local storage path for ocean data}
\usage{
get_default_ocean_path()
}
\value{
Full path to the default flow data CSV
}
\description{
Get default local storage path for ocean data
}
