% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_funs.R
\name{get_usace_flow_temp_data}
\alias{get_usace_flow_temp_data}
\title{ean daily flow (kcfs) at Bonneville from US Army Corps of Engineers}
\usage{
get_usace_flow_temp_data(forecastdate, startdate, dam = "BON")
}
\arguments{
\item{dam}{}
}
\description{
ean daily flow (kcfs) at Bonneville from US Army Corps of Engineers
}
