% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inseasonfor-package.R
\docType{package}
\name{Inseasonfor-package}
\alias{Inseasonfor}
\alias{Inseasonfor-package}
\title{Inseasonfor: Forecasts Chinook returns to the Columbia River in season}
\description{
This package forecasts Chinook returns to the Columbia River based on counts form the Bonneville Dam fish windows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/marksorel8/Inseasonfor}
  \item Report bugs at \url{https://github.com/marksorel8/Inseasonfor/issues}
}

}
\author{
\strong{Maintainer}: Mark Sorel \email{mark.sorel@dfw.wa.gov} (\href{https://orcid.org/0000-0001-5073-8052}{ORCID})

}
\keyword{internal}
