% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Joint_likelihood_RTMB.R
\name{Inseasonfor}
\alias{Inseasonfor}
\title{likelihood function for joint likelihood model in RTMB}
\usage{
Inseasonfor(data_list)
}
\arguments{
\item{parms}{}
}
\description{
likelihood function for joint likelihood model in RTMB
}
