% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Joint_likelihood_RTMB2.R
\name{Inseasonfor2}
\alias{Inseasonfor2}
\title{likelihood function for joint likelihood model in RTMB}
\usage{
Inseasonfor2(data_list)
}
\arguments{
\item{parms}{}
}
\description{
likelihood function for joint likelihood model in RTMB
}
