% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Joint_likelihood_RTMB.R
\name{joint_likelihood_results}
\alias{joint_likelihood_results}
\title{joint_likelihood_results}
\usage{
joint_likelihood_results(
  forecastdate,
  dat,
  forecast = 122500,
  forecast_log_sd = 0.28,
  joint_like_data_file = "Joint_like_res.csv"
)
}
\arguments{
\item{joint_like_data_file}{}
}
\description{
joint_likelihood_results
}
