% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Joint_likelihood_RTMB.R
\name{make_joint_like_params_fun}
\alias{make_joint_like_params_fun}
\title{make hoint likelihood parameters}
\usage{
make_joint_like_params_fun(mod_data)
}
\arguments{
\item{dat}{}
}
\description{
make hoint likelihood parameters
}
