% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Joint_likelihood_RTMB2.R
\name{make_joint_like_params_fun2}
\alias{make_joint_like_params_fun2}
\title{make joint likelihood parameters}
\usage{
make_joint_like_params_fun2(mod_data)
}
\arguments{
\item{dat}{}
}
\description{
make joint likelihood parameters
}
