% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Joint_likelihood_RTMB.R
\name{make_joint_likelihood_dat}
\alias{make_joint_likelihood_dat}
\title{joint likelihhod data constructor}
\usage{
make_joint_likelihood_dat(dat, preseason_forecast, preseason_forecast_log_sd)
}
\arguments{
\item{preseason_forecast_log_sd}{}
}
\description{
joint likelihhod data constructor
}
