% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Joint_likelihood_RTMB2.R
\name{make_joint_likelihood_dat2}
\alias{make_joint_likelihood_dat2}
\title{joint likelihhod data constructor2}
\usage{
make_joint_likelihood_dat2(dat, forecast_season)
}
\arguments{
\item{preseason_forecast_log_sd}{}
}
\description{
joint likelihhod data constructor2
}
