% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_funs.R
\name{mod_results}
\alias{mod_results}
\title{fit three models and save results to csv.}
\usage{
mod_results(
  pred_date,
  Count_dat = Bon_cnts,
  River_dat = flow_temp_dat,
  Ocean_dat = ocean_cov,
  Bon_ch_year = Bon_ch_year,
  write_local = FALSE
)
}
\arguments{
\item{joint_like_data_file}{}
}
\description{
fit three models and save results to csv.
}
