% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bon.R
\name{plot_bon}
\alias{plot_bon}
\title{plot bon}
\usage{
plot_bon(dat, forecast = 114867)
}
\arguments{
\item{dat}{dataframe containing bonneville counts, proportions of total run passed, and forecasts}

\item{forecast}{current best forecast of total season passage}
}
\value{
ggplot of observed and predicted daily counts
}
\description{
plot bon
}
\examples{
dat<-bon_ch()
plot_bon(dat,forecast =115000)

}
