% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_page.R
\name{render_and_deploy}
\alias{render_and_deploy}
\title{Render and deploy your site to GitHub Pages}
\usage{
render_and_deploy(..., commit_message = "Update GitHub Pages site")
}
\arguments{
\item{...}{Additional arguments passed to \code{render_page_fun()} (e.g., \code{params}).}

\item{commit_message}{A message to use for the Git commit.}
}
\value{
Invisibly returns the path to the deployed \code{index.html}.
}
\description{
This convenience function runs \code{render_page_fun()} and then \code{deploy_site()}.
}
