% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_funs.R
\name{bon_dat_fun}
\alias{bon_dat_fun}
\title{bon_dat_fun}
\usage{
bon_dat_fun(
  pred_date = NULL,
  count_file = NULL,
  url = "https://www.fpc.org/adults/R_coeadultcount_runsum"
)
}
\arguments{
\item{pred_date}{the last date of counts for which predictions are to be based}

\item{count_file}{the file where fish counts are stored}

\item{url}{the fishpassage center url where the data are pulled from.}
}
\value{
tibble with daily counts, proportions of runs complete, and forecasts
}
\description{
Reads window count data from fish passage center.
}
\examples{

bon_dat_fun()

}
