% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_funs.R
\name{mod_results}
\alias{mod_results}
\title{fit three models and save results to csv.}
\usage{
mod_results(
  forecastdate,
  Count_dat = Bon_cnts,
  River_dat = flow_temp_dat,
  Ocean_dat = ocean_cov,
  forecast = 122500,
  forecast_log_sd = 0.28,
  mod_result_file = NULL
)
}
\arguments{
\item{joint_like_data_file}{}
}
\description{
fit three models and save results to csv.
}
