% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_page.R
\name{render_page_fun}
\alias{render_page_fun}
\title{render_page_fun}
\usage{
render_page_fun(output_file = NULL, output_dir = NULL, params = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
render_page_fun
}
